requirements.txt
# SAM (Segment Anything Model) by FAIR - batch mask extraction
# Python >= 3.8 recommended; PyTorch with CUDA for GPU

# Core SAM - install from official repo first:
#   pip install git+https://github.com/facebookresearch/segment-anything.git

torch>=1.13.0
torchvision>=0.14.0
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.60.0
scipy>=1.7.0
matplotlib>=3.5.0

# Optional: for coco_rle output mode
# pycocotools>=2.0.6
